app/embedding_generator/requirements.txt
fastapi==0.103.2
langchain==0.0.316
pypdf==3.16.4
weaviate_client==3.24.2
uvicorn==0.23.2
python-multipart
sentence-transformers
